Move this View out ( the hosting class is already too long with ~4000 Lines ..)
