+++ k9mail/src/main/res/layout/message_compose.xml
@@ -139,7 +139,7 @@
                     <LinearLayout
                         android:orientation="horizontal"
                         android:layout_gravity="center_vertical"
-                        android:layout_width="wrap_content"
+                       android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:layout_weight="1">
 
@@ -217,7 +217,7 @@
             <!-- We have to use "wrap_content" (not "0dip") for "layout_height", otherwise the
                  EditText won't properly grow in height while the user is typing the message -->
             <view
-                class="com.fsck.k9.activity.MessageCompose$EolConvertingEditText"
+                class="com.fsck.k9.ui.EolConvertingEditText"
                 android:id="@+id/message_content"
                 android:layout_width="fill_parent"
                 android:layout_height="wrap_content"
@@ -230,7 +230,7 @@
                 android:textAppearance="?android:attr/textAppearanceMedium" />
 
             <view
-                class="com.fsck.k9.activity.MessageCompose$EolConvertingEditText"
+                class="com.fsck.k9.ui.EolConvertingEditText"
                 android:id="@+id/upper_signature"
                 android:layout_width="fill_parent"
                 android:layout_height="wrap_content"
@@ -256,7 +256,7 @@
                 android:layout_height="wrap_content">
 
                 <view
-                    class="com.fsck.k9.activity.MessageCompose$EolConvertingEditText"
+                    class="com.fsck.k9.ui.EolConvertingEditText"
                     android:id="@+id/quoted_text"
                     android:layout_width="fill_parent"
                     android:layout_height="wrap_content"
@@ -297,7 +297,7 @@
             </RelativeLayout>
 
             <view
-                class="com.fsck.k9.activity.MessageCompose$EolConvertingEditText"
+                class="com.fsck.k9.ui.EolConvertingEditText"
                 android:id="@+id/lower_signature"
                 android:layout_width="fill_parent"
                 android:layout_height="wrap_content"
